Switch file sizes to Long
The int type in Java can only have 2147483647 max, which is just about 2GB. Since downloads can get bigger than this, make file sizes in Long type. This will allow us to have downloads at more than 9000 petabytes (if ever that's possible now lmao).

+++ app/src/main/java/de/dlyt/yanndroid/freshapp/utils/Addon.java
@@ -20,7 +20,7 @@ public class Addon {
     private String mTitle;
     private String mDesc;
     private String mUpdatedOn;
-    private int mFilesize;
+    private long mFilesize;
     private String mDownloadLink;
     private int mId;
 
@@ -102,11 +102,11 @@ public class Addon {
         mDownloadLink = input;
     }
 
-    public int getFilesize() {
+    public long getFilesize() {
         return mFilesize;
     }
 
-    public void setFilesize(int input) {
+    public void setFilesize(long input) {
         mFilesize = input;
     }
